+++ src/pages/Create/ProfessionalSummary.jsx
@@ -7,7 +7,7 @@ import RedirectIfSumbitted from "hoc/RedirectIfSumbitted";
 import FormLayout from "components/Form/FormLayout";
 import TextArea from "components/Form/TextArea";
 
-import { populateResume } from "utils/resume";
+import { populateResume } from "services/resume";
 
 import { summaryIds } from "constants/formIds";
 import { CREATE } from "constants/routes";
